# media-worker-server/src/media-worker-server.js
import express from 'express'
import bodyParser from 'body-parser'
import dotenv from 'dotenv-defaults'
import http from 'http'
import * as mediasoup from 'mediasoup'
import { MEDIA_CODECS, WORKER_SETTINGS } from './constants'
import logger from 'morgan'

dotenv.config()

const transports = new Map(),
  producers = new Map(),
  consumers = new Map()

let app, httpServer, worker, router

main()

async function main() {
  await startMediasoup()
  startWebserver()

  setInterval(() => {
    console.log(
      'transports:',
      [...transports.values()].map((t) => [t.id, t.appData.clientDirection]),
      'producers:',
      [...producers.values()].map((p) => p.id),
      'consumers:',
      [...consumers.values()].map((c) => c.id)
    )
  }, 5000)
}

async function startMediasoup() {
  worker = await mediasoup.createWorker(WORKER_SETTINGS)

  worker.on('died', () => {
    console.error('mediasoup worker died (this should never happen)')
    process.exit(1)
  })

  router = await worker.createRouter({ mediaCodecs: MEDIA_CODECS })
}

function startWebserver() {
  const { PORT } = process.env
  app = express()

  app.use(bodyParser.urlencoded({ extended: false }))
  app.use(bodyParser.json())
  app.use(logger('dev'))

  app.get('/rooms/:roomName', (req, res) => {
    return res.json({ routerRtpCapabilities: router.rtpCapabilities })
  })

  app.post('/rooms/:roomName/peers', (req, res) => {
    return res.json({ roomName: req.params.roomName })
  })

  app.delete('/rooms/:roomName/peers/:socketId', async (req, res) => {
    const { socketId } = req.params
    await closePeer(socketId)
    return res.json({ roomName: req.params.roomName, socketId })
  })

  app.post('/rooms/:roomName/transports', async (req, res) => {
    const { fromSocketId, direction, toSocketId } = req.body

    const transport = await createWebRtcTransport({
      fromSocketId,
      direction,
      toSocketId,
    })
    transports.set(transport.id, transport)
    const { id, iceParameters, iceCandidates, dtlsParameters } = transport
    return res.json({
      transportOptions: { id, iceParameters, iceCandidates, dtlsParameters },
    })
  })

  httpServer = http.Server(app)
  httpServer.listen(PORT, () => {
    console.log(`Listening HTTP in port ${PORT}`)
  })

  app.put('/rooms/:roomName/transports/:transportId', async (req, res) => {
    const { socketId, dtlsParameters } = req.body
    const { transportId } = req.params

    const transport = transports.get(transportId)

    if (!transport) {
      return res.status(404).json({ error: 'Transport not found' })
    } else if (transport.appData.socketId !== socketId) {
      return res.status(403).json({ error: 'Transport is forbidden' })
    }

    try {
      await transport.connect({ dtlsParameters })
    } catch (error) {
      console.error('error in connect-transport', {
        socketId,
        transportId,
        dtlsParameters,
        error,
      })
      return res.status(500).json({ error: 'Could not connect transport' })
    }

    return res.json({})
  })

  app.post(
    '/rooms/:roomName/transports/:transportId/producers',
    async (req, res) => {
      const { socketId, kind, rtpParameters, paused, appData } = req.body
      const { transportId } = req.params

      const transport = transports.get(transportId)

      if (!transport) {
        return res.status(404).json({ error: 'Transport not found' })
      } else if (transport.appData.fromSocketId !== socketId) {
        return res.status(403).json({ error: 'Transport is forbidden' })
      }

      let producer = await transport.produce({
        kind,
        rtpParameters,
        paused,
        appData: { ...appData, socketId, transportId },
      })

      producer.on('transportclose', () => {
        console.log("producer's transport closed", {
          socketId,
          producerId: producer.id,
        })
        closeProducer(producer)
      })

      if (producer.kind === 'audio') {
        // TODO: observe audio producer
      }

      producers.set(producer.id, producer)
      return res.json({ producerId: producer.id })
    }
  )

  app.post(
    `/rooms/:roomName/transports/:transportId/consumers`,
    async (req, res) => {
      const { fromSocketId, toSocketId, mediaTag, rtpCapabilities } = req.body
      const { transportId } = req.params
      const producer = Array.from(producers.values()).find(
        (p) =>
          p.appData.mediaTag === mediaTag && p.appData.socketId === toSocketId
      )

      if (!producer) {
        console.error(
          `service-side producer for ${toSocketId}:${mediaTag} not found`
        )
        return res.status(404).json({ error: 'Producer not found' })
      }

      let transport = transports.get(transportId)

      if (!transport) {
        console.error(`service-side recv transport #${transportId} not found`)
        return res.status(404).json({ error: 'Transport not found' })
      }

      if (!router.canConsume({ producerId: producer.id, rtpCapabilities })) {
        console.error(`client cannot consume ${toSocketId}:${mediaTag}`)
        return res
          .status(500)
          .json({ error: "Router can't consume from producer" })
      }

      let consumer = await transport.consume({
        producerId: producer.id,
        rtpCapabilities,
        paused: false,
        appData: { fromSocketId, toSocketId },
      })

      consumer.on('transportclose', async () => {
        console.log(`consumer's transport closed`, consumer.id)
        await closeConsumer(consumer)
      })

      consumer.on('producerclose', async () => {
        console.log(`consumer's producer closed`, consumer.id)
        await closeConsumer(consumer)
      })

      consumers.set(consumer.id, consumer)

      return res.json({
        producerId: producer.id,
        id: consumer.id,
        kind: consumer.kind,
        rtpParameters: consumer.rtpParameters,
        type: consumer.type,
        producerPaused: consumer.producerPaused,
      })
    }
  )
}

async function createWebRtcTransport({ fromSocketId, direction, toSocketId }) {
  const transport = await router.createWebRtcTransport({
    listenIps: [
      {
        ip: process.env.LISTEN_IP,
        announcedIp: process.env.ANNOUNCED_IP || null,
      },
    ],
    enableUdp: true,
    enableTcp: true,
    preferUdp: true,
    initialAvailableOutgoingBitrate: 800000,
    appData: { fromSocketId, clientDirection: direction, toSocketId },
  })

  return transport
}

async function closePeer(socketId) {
  console.log('closing peer', { socketId })
  for (const [, transport] of transports) {
    if (
      transport.appData.fromSocketId === socketId ||
      transport.appData.toSocketId === socketId
    ) {
      await closeTransport(transport)
    }
  }
}

async function closeTransport(transport) {
  try {
    console.log('closing transport', {
      transportId: transport.id,
      appData: transport.appData,
    })
    await transport.close()
    transports.delete(transport.id)
  } catch (error) {
    console.error(error)
  }
}

async function closeProducer(producer) {
  console.log('closing producer', {
    producerId: producer.id,
    appData: producer.appData,
  })

  try {
    await producer.close()
    producers.delete(producer.id)
  } catch (error) {
    console.error(error)
  }
}

async function closeConsumer(consumer) {
  console.log('closing consumer', {
    consumerId: consumer.id,
    appData: consumer.appData,
  })

  try {
    await consumer.close()
    consumers.delete(consumer.id)
  } catch (error) {
    console.error(error)
  }
}